Fix: pasar logica a repositorio
